backend/requirements.txt
fastapi
uvicorn
celery==5.3.6
redis
python-multipart
pillow
numpy
requests
slowapi
python-magic
